requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
geopy==2.3.0
